Temperature calculation added - readable by human eyes. Generic slot added to support GUI. Parameters are ID and 32 bit data. It is enough for all kinds of  messages coming from the GUI. All controllers emit signals to this slot.
TODO:
Code cleanup.
Add error message signal for GUI.
Integrate with GUI.

+++ Telemetry/Telemetry/application.h
@@ -37,7 +37,7 @@ public:
     QVector<qint8> payload;
     Package();
     Package(QByteArray data);
-    Package(quint8 ID, qint8 data);
+    Package(quint8 ID, qint32 data);
     Package(const Package& other);
 };
 
@@ -48,10 +48,10 @@ QDataStream& operator<<(QDataStream& out, const Package& package);
 class Temperatures
 {
 public:
-    qint32 temp1;
-    qint32 temp2;
-    qint32 temp3;
-    qint32 temp4;
+    double temp1;
+    double temp2;
+    double temp3;
+    double temp4;
 
     void reset();
     Temperatures& operator=(Temperatures&& other); //move assignment
